requirements.txt
streamlit
google-generativeai
requests
beautifulsoup4
duckduckgo-search
python-dotenv
PyMuPDF